refactor dpm pm quality routes
